+++ server/src/main/kotlin/endpoints/v1/api/OrganizationsEndpoint.kt
@@ -19,13 +19,24 @@
 
 package org.noelware.charted.server.endpoints.v1.api
 
+import dev.floofy.utils.exposed.asyncTransaction
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.response.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import org.jetbrains.exposed.sql.Op
+import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
+import org.jetbrains.exposed.sql.and
 import org.noelware.charted.ChartedInfo
+import org.noelware.charted.ChartedScope
 import org.noelware.charted.configuration.kotlin.dsl.Config
+import org.noelware.charted.databases.postgres.entities.OrganizationEntity
+import org.noelware.charted.databases.postgres.entities.UserEntity
+import org.noelware.charted.databases.postgres.models.Organization
+import org.noelware.charted.databases.postgres.tables.OrganizationTable
+import org.noelware.charted.databases.postgres.tables.UserTable
+import org.noelware.charted.extensions.regexp.toNameRegex
 import org.noelware.charted.modules.avatars.AvatarModule
 import org.noelware.charted.modules.helm.charts.HelmChartModule
 import org.noelware.charted.modules.storage.StorageHandler
@@ -41,6 +52,41 @@ data class MainOrganizationsResponse(
     @SerialName("docs_url")
     val docsUrl: String = "https://charts.noelware.org/docs/server/${ChartedInfo.version}/api/organizations"
 )
+
+@Serializable
+data class CreateOrganizationBody(
+    @SerialName("display_name")
+    val displayName: String? = null,
+    val owner: Long,
+    val name: String
+)
+
+@Serializable
+data class PatchOrganizationBody(
+    @SerialName("twitter_handle")
+    val twitterHandle: String? = null,
+
+    @SerialName("gravatar_email")
+    val gravatarEmail: String? = null,
+
+    @SerialName("display_name")
+    val displayName: String? = null,
+    val name: String? = null
+)
+
+/*
+    var verifiedPublisher by OrganizationTable.verifiedPublisher
+    var twitterHandle by OrganizationTable.twitterHandle
+    var gravatarEmail by OrganizationTable.gravatarEmail
+    var displayName by OrganizationTable.displayName
+    var createdAt by OrganizationTable.createdAt
+    var updatedAt by OrganizationTable.updatedAt
+    var iconHash by OrganizationTable.iconHash
+    val members by OrganizationMemberEntity referrersOn OrganizationMemberTable.organization
+    var owner by UserEntity referencedOn OrganizationTable.owner
+    var flags by OrganizationTable.flags
+    var name by OrganizationTable.name
+ */
 
 class OrganizationsEndpoint(
     private val config: Config,
@@ -167,12 +213,68 @@ class OrganizationsEndpoint(
 
     @Get("/{id}")
     suspend fun getById(call: ApplicationCall) {
-        call.respond(HttpStatusCode.NotImplemented)
+        val organization = call.getOrganizationById() ?: return
+        call.respond(HttpStatusCode.OK, ApiResponse.ok(Organization.fromEntity(organization)))
     }
 
     @Get("/{owner}/{name}")
     suspend fun getByOwnerAndName(call: ApplicationCall) {
-        call.respond(HttpStatusCode.NotImplemented)
+        val owner = call.parameters["owner"]!!
+        val name = call.parameters["name"]!!
+        val ownerId = when {
+            owner.toLongOrNull() != null -> owner.toLong()
+            owner.toNameRegex(false).matches() -> asyncTransaction(ChartedScope) {
+                UserEntity.find { UserTable.username eq name }.firstOrNull()?.id?.value
+            }
+
+            else -> {
+                call.respond(
+                    HttpStatusCode.BadRequest,
+                    ApiResponse.err(
+                        "UNKNOWN_ENTITY",
+                        "Unable to determine if [owner] path param provided is by ID or name, provided [$owner]"
+                    )
+                )
+
+                null
+            }
+        } ?: return run {
+            if (!call.isHandled) {
+                call.respond(
+                    HttpStatusCode.NotFound,
+                    ApiResponse.err(
+                        "UNKNOWN_USER",
+                        "User with username [$owner] was not found."
+                    )
+                )
+            }
+        }
+
+        val orgSearchQuery: Op<Boolean> = if (name.toLongOrNull() != null) {
+            OrganizationTable.id eq name.toLong()
+        } else if (name.toNameRegex(false, 32).matches()) {
+            OrganizationTable.name eq name
+        } else {
+            return call.respond(
+                HttpStatusCode.BadRequest,
+                ApiResponse.err(
+                    "UNKNOWN_ENTITY", "Unable to determine if [organization name] path param provided is a snowflake or name, provided [$name]"
+                )
+            )
+        }
+
+        val organization = asyncTransaction(ChartedScope) {
+            OrganizationEntity.find {
+                orgSearchQuery and (OrganizationTable.owner eq ownerId)
+            }.firstOrNull()?.let { entity -> Organization.fromEntity(entity) }
+        } ?: return call.respond(
+            HttpStatusCode.NotFound,
+            ApiResponse.err(
+                "UNKNOWN_ORGANIZATION", "Unable to find organization by owner/name [$owner/$name]"
+            )
+        )
+
+        call.respond(HttpStatusCode.OK, ApiResponse.ok(organization))
     }
 
     @Put
@@ -288,5 +390,64 @@ class OrganizationsEndpoint(
     @Delete("/{idOrName}/webhooks/{webhookId}/events/{eventId}")
     suspend fun deleteWebhookEvent(call: ApplicationCall) {
         call.respond(HttpStatusCode.NotImplemented)
+    }
+
+    private suspend fun ApplicationCall.getOrganization(by: String, query: Op<Boolean>): OrganizationEntity? = asyncTransaction(ChartedScope) {
+        OrganizationEntity.find(query).firstOrNull()
+    } ?: run {
+        respond(
+            HttpStatusCode.NotFound,
+            ApiResponse.err(
+                "UNKNOWN_ORGANIZATION", "Unable to find organization by $by"
+            )
+        )
+
+        null
+    }
+
+    private suspend fun ApplicationCall.getOrganizationById(): OrganizationEntity? {
+        val id = parameters["id"]?.toLongOrNull()
+            ?: return run {
+                respond(
+                    HttpStatusCode.BadRequest,
+                    ApiResponse.err(
+                        "INVALID_ORGANIZATION_ID",
+                        "Provided organization ID was not a valid snowflake"
+                    )
+                )
+
+                null
+            }
+
+        return getOrganization("id path parameter [$id]", OrganizationTable.id eq id)
+    }
+
+    private suspend fun ApplicationCall.getOrganizationByIdOrName(): OrganizationEntity? {
+        val idOrName = parameters["idOrName"] ?: return run {
+            respond(
+                HttpStatusCode.NotFound,
+                ApiResponse.err(
+                    "MISSING_PARAMETER", "Missing [idOrName] path parameter"
+                )
+            )
+
+            null
+        }
+
+        return when {
+            idOrName.toLongOrNull() != null -> getOrganization("id path parameter [$idOrName]", OrganizationTable.id eq idOrName.toLong())
+            idOrName.toNameRegex(false, 24).matches() -> getOrganization("org name path parameter [$idOrName]", OrganizationTable.name eq idOrName)
+            else -> {
+                respond(
+                    HttpStatusCode.BadRequest,
+                    ApiResponse.err(
+                        "UNKNOWN_ENTITY",
+                        "Unable to determine if [idOrName] provided is by ID or name, provided [$idOrName]"
+                    )
+                )
+
+                null
+            }
+        }
     }
 }